refactor(cli): move splitKeyValue to output.ts so location matches its convention status (#165)
CLAUDE.md § "Conventions" names `splitKeyValue` as a durable
cross-feature CLI primitive, but the function itself was private to
`src/cli/traversals.ts`. A new CLI surface introducing a `key=value`
flag outside traversals would either duplicate the split-on-first-`=`
+ empty-key guard or reach into a domain-specific handler file for
it — neither reflects the "single primitive" the convention stakes
out.

Relocate the helper to `src/cli/output.ts` alongside `parseIntArg`,
which already hosts the cross-feature CLI primitives. Update the
one existing caller (`traversals.ts`) to import, and update the
CLAUDE.md convention paragraph to point at the new home.

No behavior change — same signature, same guard, same error codes.
